During make install copy fio_py for Python 3
The name of the Python module can differ from "fio_py.so", if it is created with a Python version different from default. Makefile was updated to copy any file named "fio_py*.so" to the $(FIO_INSTALL_DIR)/lib directory. Manual renaming of this file to fio_py.so is required.